Split out rec-main.cc
